+++ src/OpenTelemetry.Instrumentation.Http/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * .NET Framework - fix description for `http.client.request.duration` metric.
-  ([#1538](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1538))
+  ([#5234](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5234))
 
 ## 1.7.0